Updates
Fixed and sped up glmmMixture function. It isn’t working as well as it
used to, or as well as the multivariate function. But I’m pretty sure
it’s algorithmically correct now.

Edited the multivariate function so the treatment variable can be a
factor. An open issue is that it has to be called ‘treatment’ I think I
know how to fix that though.

Did some analysis for the malaria dataset.