Some extreme case adjustments + Malaria analysis
Alternative initialization for random effects when lme4 fails.

A linear combination of positive definite matrices is positive
definite? Not in R land! (so covariance is no longer updated as an
average.

assignmentList is now named, so the ising model fit now includes
subpopulation names.

Working on jackknife standard error computation for Malaria dataset.