Added option to replicate dataset
This is equivalent to sampling several posteriors per subject per
iteration and averaging them in the fitting process. This is the
standard recommendation procedure for stochastic EM. So far we have
been averaging over many iterations instead of averaging within
iterations. Averaging within iteration has an added benefit of making
the algorithm more stable when the sample size is small (such as in the
case of the malaria dataset). The posterior probabilities need to be
averaged once the algorithm converges.