Updates to pre-assignments and plotting of Graphs
1. Instead of hard 0/1 decisions, pre-assignment can now be used to
guide the fitting algorithm through the estimation process by providing
a vector of coefficients  with which the probability of
response/non-response will be multiplied in the Gibbs sampler. This
approach leads to better fitting while retaining more of the
information contained in the magnitude of the observed proportions and
dependencies on other cell-subsets.

2. Implemented a function for plotting the raw graph used in fitting
the model, accessible through the S3 plotting method. The thresholding
of edges is based on their quantile.

3. New analysis of Malaria dataset where the only time varying
component is the stimulation effect.