Major update: isingWprior
Implemented an option for fitting the ising model is a `moderated’ way,
to prevent the Ising model from estimating very high (or low) response
probabilities when it is completely uncalled for. Works by fixing the
threshold of the ising model to ensure a desired response level and
estimating the diagonal elements of the Ising matrix using glmnet.